+++ src/sender/udp.rs
@@ -17,7 +17,8 @@ use std::net::ToSocketAddrs;
 use std::net::UdpSocket;
 
 use crate::format::SyslogContext;
-use crate::sender::internal::impl_datagram_syslog_sender;
+use crate::SDElement;
+use crate::Severity;
 
 /// Create a UDP sender that sends messages to the well-known port (514).
 ///
@@ -60,6 +61,30 @@ impl UdpSender {
     pub fn mut_context(&mut self) -> &mut SyslogContext {
         &mut self.context
     }
-}
 
-impl_datagram_syslog_sender!(UdpSender, socket);
+    /// Send a message with the given severity as defined in RFC-3164.
+    pub fn send_rfc3164<M: std::fmt::Display>(
+        &mut self,
+        severity: Severity,
+        message: M,
+    ) -> io::Result<()> {
+        let message = self.context.format_rfc3164(severity, Some(message));
+        self.socket.send(message.to_string().as_bytes())?;
+        Ok(())
+    }
+
+    /// Send a message with the given severity as defined in RFC-5424.
+    pub fn send_rfc5424<S: Into<String>, M: std::fmt::Display>(
+        &mut self,
+        severity: Severity,
+        msgid: Option<S>,
+        elements: Vec<SDElement>,
+        message: M,
+    ) -> io::Result<()> {
+        let message = self
+            .context
+            .format_rfc5424(severity, msgid, elements, Some(message));
+        self.socket.send(message.to_string().as_bytes())?;
+        Ok(())
+    }
+}
